refactor: 점수 VO 디렉터리 변경
