+++ src/main/java/blackjack/domain/card/Cards.java
@@ -1,6 +1,5 @@
 package blackjack.domain.card;
 
-import blackjack.domain.participant.Score;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
